Added get parameter redirect functionality on login button click.
I added functionality to the login button with the ID "loginBtn". When the button is clicked, the event is captured, and the default form submission behavior is prevented using event.preventDefault(). Then, the code retrieves the value of the "redirected" parameter from the current URL using URLSearchParams, which allows extracting query parameters. If a "redirected" parameter exists, the value is decoded using decodeURIComponent(). Finally, the user is redirected to the specified URL using window.location.href.