Redirects the user to a specified URL using PHP.
I added a URL parameter is obtained from the GET request using $_GET['url']. If the URL parameter is not empty, the code performs a redirection to the specified URL using the header() function. The header() function sets the HTTP header "Location" to the provided URL, instructing the browser to redirect the user to that URL. After setting the header, the exit statement is called to halt further script execution.

This code snippet allows for dynamic redirection based on the value passed in the "url" parameter of the GET request. It can be used to redirect users to different pages or external websites based on the provided URL. It provides flexibility in handling redirects within PHP applications.

This code is integrated with jwellery.html, fashion.html and electronic.html to provide path fragment redirection.