Added login functionality and Path fragment & Get parameter vulnerability
I added login functionality and get & path vulnerability. When the user submits the login form, the code captures the entered username and password. It then establishes a connection to the database, queries the "user" table to check if the provided username and password match an entry in the database, and displays appropriate alerts based on the authentication result.

If the login is successful, the code triggers an alert message and checks for a "redirected" parameter in the URL. If the parameter exists, the user is redirected to the specified URL. Otherwise, they are redirected to a default page, such as "../Html/photo_uploader.html".

On the HTML side, a login form is displayed, featuring fields for entering the username and password. The form is validated to ensure that both fields are filled out. The form submission triggers the PHP code for authentication and processing.